Made requested changes see comment in merge request
