+++ Assets/Concept 2/Objects/Prefabs/Crafter.prefab
@@ -1705,8 +1705,8 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 2dbbdefbebefd544cadf419b51b3a32d, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  line1: {fileID: 3481057997310183538}
-  line2: {fileID: 8138445477234787703}
+  line1: {fileID: 6018000140153687598}
+  line2: {fileID: 7876169785252658364}
   intersection: {x: 0, y: 0}
 --- !u!1 &5674705329855749110
 GameObject:
