+++ Assets/Scripts/Crafter/InterceptionCalculator.cs
@@ -4,8 +4,8 @@ namespace Crafter
 {
     public class InterceptionCalculator : MonoBehaviour
     {
-        [SerializeField] private GameObject line1;
-        [SerializeField] private GameObject line2;
+        [SerializeField] private FunctionLineController line1;
+        [SerializeField] private FunctionLineController line2;
 
         public Vector2 intersection;
 
@@ -23,29 +23,7 @@ namespace Crafter
             _a2 = line2.GetComponent<FunctionLineController>().A;
             _b2 = line2.transform.position.y;
 
-            intersection = FindIntersection(_a1, _b1, _a2, _b2);
-        }
-
-        /// <summary>
-        /// Finds intersection between 2 lines
-        /// </summary>
-        /// <param name="a1">Slope of line 1</param>
-        /// <param name="b1">Intercept of line 1</param>
-        /// <param name="a2">Slope of line 2</param>
-        /// <param name="b2">Intercept of line 2</param>
-        /// <returns>The correct interception for 2 lines as vector2</returns>
-        private Vector2 FindIntersection(float a1, float b1, float a2, float b2)
-        {
-            if (a1 == a2)
-            {
-                Debug.Log("Lines are parallel and do not intersect.");
-                return Vector2.zero;
-            }
-
-            var xIntercept = (b2 - b1) / (a1 - a2);
-            var yIntercept = a1 * xIntercept + b1;
-
-            return new Vector2(xIntercept, yIntercept);
+            intersection = Vector2Extension.FindIntersection(_a1, _b1, _a2, _b2);
         }
     }
 }
